+++ lib/tests/helpers.test.js
@@ -4,11 +4,11 @@ import {
   rehydrateFormResponses,
   extractFormData,
   submitToAPI,
-} from "../integration/helpers";
+} from "../helpers";
 
 import { logMessage } from "@lib/logger";
 import TestForm from "../../__fixtures__/testData";
-import { getFormByID } from "../integration/helpers";
+import { getFormByID } from "../helpers";
 
 jest.mock("axios");
