+++ pages/api/submit.ts
@@ -2,7 +2,7 @@ import { NotifyClient } from "notifications-node-client";
 import type { NextApiRequest, NextApiResponse } from "next";
 import { LambdaClient, InvokeCommand } from "@aws-sdk/client-lambda";
 import convertMessage from "@lib/markdown";
-import { rehydrateFormResponses } from "@lib/integration/helpers";
+import { rehydrateFormResponses } from "@lib/helpers";
 import { getFormByID, getSubmissionByID } from "@lib/integration/crud";
 import { logMessage } from "@lib/logger";
 import { checkOne } from "@lib/flags";
@@ -279,7 +279,7 @@ const processFormData = async (
       }`
     );
 
-    const form = await getFormByID(reqFields.formID);
+    const form = await getFormByID(reqFields.formID as string);
 
     if (!form) {
       return res.status(400).json({ error: "No form could be found with that ID" });
